Fix compilation on gcc
