+++ src/ifcgeom/IfcGeomFilter.h
@@ -144,7 +144,7 @@ namespace IfcGeom
         {
 #ifndef NDEBUG
             for (arg_map_t::const_iterator it = args.begin(); it != args.end(); ++it) {
-                IfcWrite::IfcWritableEntity dummy(it->first);
+                IfcEntityInstanceData dummy(it->first);
                 IfcUtil::IfcBaseClass* base = IfcSchema::SchemaEntity(&dummy);
                 assert(it->second < base->getArgumentCount() && "Argument index out of bounds");
                 assert(base->getArgumentType(it->second) == IfcUtil::Argument_STRING && "Argument type not string");
