+++ src/ifcparse/Argument.h
@@ -52,34 +52,6 @@ namespace IfcParse {
 }
 namespace IfcUtil {
 	class IfcBaseClass;
-}
-
-namespace IfcUtil {
-	enum ArgumentType {
-		Argument_NULL,
-		Argument_DERIVED,
-		Argument_INT,
-		Argument_BOOL,
-		Argument_DOUBLE,
-		Argument_STRING,
-		Argument_BINARY,
-		Argument_ENUMERATION, 
-		Argument_ENTITY_INSTANCE,
-
-		Argument_EMPTY_AGGREGATE,
-		Argument_AGGREGATE_OF_INT,
-		Argument_AGGREGATE_OF_DOUBLE, 
-		Argument_AGGREGATE_OF_STRING,
-		Argument_AGGREGATE_OF_BINARY, 
-		Argument_AGGREGATE_OF_ENTITY_INSTANCE,
-	
-		Argument_AGGREGATE_OF_EMPTY_AGGREGATE,
-		Argument_AGGREGATE_OF_AGGREGATE_OF_INT,
-		Argument_AGGREGATE_OF_AGGREGATE_OF_DOUBLE,
-		Argument_AGGREGATE_OF_AGGREGATE_OF_ENTITY_INSTANCE, 
-
-		Argument_UNKNOWN
-	};
 
     IFC_PARSE_API const char* ArgumentTypeToString(ArgumentType argument_type);
